Style tweaks
Fix some minor style issues in vtkMaptkCameraRepresentation and
generally clean up the code. Fix wrong type of DisplaySkip.

+++ gui/vtkMaptkCameraRepresentation.h
@@ -111,7 +111,7 @@ private:
 
   double ActiveCameraRepLength;
   double NonActiveCameraRepLength;
-  double DisplaySkip;
+  int DisplaySkip;
 
   vtkCamera* ActiveCamera;
